+++ intents/fulfillment.py
@@ -148,13 +148,14 @@ def run_dev_server(connector: Connector, host: str='', port: str=8000):
             content_len = int(self.headers.get('Content-Length'))
             post_body = self.rfile.read(content_len)
             post_body = json.loads(post_body)
-            logger.info("POST REQUEST BODY: %s", jsondict(post_body))
+            logger.info("POST Request reveived")
+            logger.debug("POST REQUEST BODY: %s", jsondict(post_body))
 
             fulfillment_request = FulfillmentRequest(
                 body=post_body
             )
             result = connector.fulfill(fulfillment_request)
-            logger.info("POST RESPONSE: %s", jsondict(result))
+            logger.debug("POST RESPONSE: %s", jsondict(result))
             result = json.dumps(result)
             
             self.send_response(http.server.HTTPStatus.OK)